@@ -1,14 +1,14 @@
 
 object Version {
-    const val kotlin = "1.5.21"
+    const val kotlin = "1.6.0"
     const val snapshot = "7055167"
 
     const val fragmentKtx = "1.4.0-alpha04"
 
-    const val compose = "1.0.1"
-    const val composeNavigation = "1.0.0-alpha10"
+    const val compose = "1.2.0-alpha02"
+    const val composeNavigation = "2.5.0-alpha01"
 
-    const val hilt = "2.37"
+    const val hilt = "2.40.5"
     const val hiltNavigation = "1.0.0-alpha03"
     const val hiltNavigationFragment = "1.0.0"
 
